+++ GRDB/Record/FetchableRecord+Decodable.swift
@@ -147,6 +147,7 @@ private struct _RowDecoder<R: FetchableRecord>: Decoder {
         
         lazy var allKeys: [Key] = {
             let row = decoder.row
+            // TODO: test when _columnForKey is not nil
             var keys = _columnForKey.map { Set($0.keys) } ?? Set(row.columnNames)
             keys.formUnion(row.scopesTree.names)
             keys.formUnion(row.prefetchedRows.keys)
@@ -174,15 +175,25 @@ private struct _RowDecoder<R: FetchableRecord>: Decoder {
         
         func decodeNil(forKey key: Key) throws -> Bool {
             let row = decoder.row
-            if let column = try? decodeColumn(forKey: key), row[column] != nil {
-                return false
+            
+            // Column?
+            if let column = try? decodeColumn(forKey: key),
+               let index = row.index(forColumn: column)
+            {
+                return row.hasNull(atIndex: index)
             }
-            if row.scopesTree[key.stringValue] != nil {
-                return false
+            
+            // Scope?
+            if let scopedRow = row.scopesTree[key.stringValue] {
+                return scopedRow.containsNonNullValue == false
             }
+            
+            // Prefetched Rows?
             if row.prefetchedRows[key.stringValue] != nil {
                 return false
             }
+            
+            // Unknown key
             return true
         }
         
@@ -376,7 +387,37 @@ private struct _RowDecoder<R: FetchableRecord>: Decoder {
         func nestedContainer<NestedKey>(keyedBy type: NestedKey.Type, forKey key: Key)
         throws -> KeyedDecodingContainer<NestedKey> where NestedKey: CodingKey
         {
-            fatalError("not implemented")
+            let row = decoder.row
+            
+            // Column?
+            if let column = try? decodeColumn(forKey: key),
+               row.index(forColumn: column) != nil
+            {
+                // We need a JSON container, but how do we create one?
+                throw DecodingError.typeMismatch(
+                    KeyedDecodingContainer<NestedKey>.self,
+                    DecodingError.Context(
+                        codingPath: codingPath,
+                        debugDescription: """
+                            not implemented: building a nested JSON container for the column '\(column)'
+                            """))
+            }
+            
+            // Scope?
+            if let scopedRow = row.scopesTree[key.stringValue] {
+                return KeyedDecodingContainer(KeyedContainer<NestedKey>(decoder: _RowDecoder(
+                    row: scopedRow,
+                    codingPath: codingPath + [key],
+                    columnDecodingStrategy: decoder.columnDecodingStrategy)))
+            }
+            
+            // Don't look for prefetched rows: those need a unkeyed container.
+            
+            throw DecodingError.typeMismatch(
+                KeyedDecodingContainer<NestedKey>.self,
+                DecodingError.Context(
+                    codingPath: codingPath,
+                    debugDescription: "No keyed container found for key '\(key)'"))
         }
         
         func nestedUnkeyedContainer(forKey key: Key) throws -> UnkeyedDecodingContainer {